fix(ha): scope HA capability/debt state to its real owner (#5718 C-HA)
Implements the five items the #5718 audit confirmed still unimplemented
on master; C01b/C01c shipped earlier in PR #6376. Each was re-verified
against origin/master ad9591177 before any code was written, per the
issue's own method note that a naive item-ID grep hits docs/reviews/**
and _Log.md for 100% of items and falsely reads as "all fixed".

C01a — the peer heartbeat-ACK capability was process-sticky.
SessionSync.peerHeartbeatAckEver latches when the connected peer replies
syncMsgHeartbeatAck, and that latch is what switches two paths from
"assume healthy" to "enforce": the receiveLoop missed-heartbeat teardown
(two read deadlines with no traffic closes the connection) and
PeerHealthy()'s silence window, which computeUserspaceTransferReadiness
consults before allowing a manual failover. It was never cleared, so it
outlived the peer incarnation that earned it. On a peer DOWNGRADE — new
build acks, peer rolls back to a build that never acks — the stale latch
turned a healthy old peer into connection churn every two read deadlines
plus a permanent "session sync disconnected" failover block. It is now
cleared in handleDisconnect's FULL-disconnect block, directly beside
clockSynced: the sibling peer-incarnation capability that was already
reset correctly, same reason and same place. It is deliberately NOT
cleared on a partial disconnect — one fabric link dropping while the
other holds is the same peer process, and resetting there would disarm
both enforcement paths on every link blip.

A6-b01-C1 — the worker clamp cast before it clamped. heartbeatZeroSlots
did uint32(maxInt(workers, 1)) BEFORE comparing against
mapCap/heartbeatSlotsPerWorker, so the narrowing happened before either
clamp could see the value. workers = 1<<32 narrowed to 0, passed under
the high clamp, and returned 0 — zero-initialising nothing and leaving
every worker on stale heartbeat data; 1<<32+5 narrowed to 5. workers is
a min-only schema leaf with no upper bound in deriveUserspaceConfig, so
both are reachable from config. Both clamps now run in int space and the
cast happens on a value already proven to be within [1, maxW]. The
existing #4572 test stopped at 999999999, below the narrowing boundary,
so it could not observe either case.

A7-b01-C001 — networkd Clear forgot its activation debt. Removing the
managed files deactivates nothing until networkctl reload lands. Clear
recorded no reloadPending on reload failure and returned early on an
empty glob, so the second Clear found the files already gone and
reported a success it had not achieved while the removed addresses,
VRFs, bonds and renames stayed live. Clear now mirrors Apply's #4954
debt contract on both halves: record the debt on failure, and on an
empty glob re-run the idempotent reload when debt is outstanding rather
than assuming "no files" means "nothing to activate".

A7-b02-C01 — the partial test-manager Close panicked. Manager.Close()
called m.tunnel.stopAll() unguarded while NewManagerWithRuleOpsForTest
and NewManagerWithRouteListerForTest leave m.tunnel nil and their doc
comments promised "Close nil-guards it" — true only of nlHandle.
stopAll takes t.mu immediately, so the documented `defer m.Close()`
nil-dereferenced. The tunnel domain is now guarded exactly as nlHandle
is, and the two stale comments are corrected.

D-A6-b00-C1 — the user app-set NAT guard false-passed. A membership-only
`if !want[pp]` check behind a length check is satisfied by any multiset
whose members all appear in want, so two duplicate tcp/4444 terms passed
while a missing udp/5555 went undetected: the guard could not observe
the regression it exists to catch. Rewritten to consume with delete()
and assert the want set drains, matching the correct sibling test in the
same file, plus the per-term shape assertions that keep Ports[0] from
being an unchecked index.

Validation: every guard was mutation-proved — reverted individually,
watched the specific test go RED from an assertion (or, for the Close
guard, the real nil dereference at routing.go:91), restored with touch,
watched it go GREEN. go vet was clean under every mutation, so no RED
was a build break, and build+vet plus free disk were confirmed before
any mutation so no RED could be a stale-cache or full-disk artifact.
C01a was mutated in BOTH directions: dropping the reset fails the
incarnation test, hoisting it to every disconnect fails the
partial-disconnect scope control, so the guard is pinned to the peer
incarnation and is neither narrower nor broader than its claim. The
D-A6-b00-C1 proof is two-sided — with the duplicate-plus-missing
regression injected, the strengthened guard FAILS and the pre-fix guard
PASSES. go test -race ./pkg/cluster/... ./pkg/vrrp/... is clean and the
full go test ./... is clean (59 packages, 0 failures).

No cluster/incus/smoke tooling was run: the issue notes that promoting
any of these to material needs test-failover, which the campaign lead
schedules.

Advances #5718.

+++ pkg/dataplane/userspace/maps_sync.go
@@ -1828,12 +1828,23 @@ const heartbeatSlotsPerWorker = 2 * 16
 // iterations that hang the apply for hours (#4572). Clamping the worker
 // count to mapCap/heartbeatSlotsPerWorker before the multiply keeps the
 // returned slot count <= mapCap and never wraps.
+// #5718 A6-b01-C1: BOTH clamps run in int space, BEFORE the uint32 cast.
+// Casting first (the pre-#5718 shape, `w := uint32(maxInt(workers, 1))`) let a
+// worker count above the uint32 range narrow into the clamp's blind spot and
+// defeat the very bounds this function exists to enforce: `workers` is an int
+// from a min-only schema leaf, so `1<<32` narrows to 0 — under maxW, so the
+// high clamp passes it through — and the function returns 0 slots, zeroing
+// NOTHING and leaving stale heartbeat entries live for every worker. `1<<32+5`
+// narrows to 5 and silently zeroes 5 workers' slots no matter how large maxW
+// is. Clamping in int space keeps the low clamp (>=1) and the high clamp
+// (<=maxW) both binding for every int input, and the cast then happens on a
+// value already proven to be within [1, maxW].
 func heartbeatZeroSlots(workers int, mapCap uint32) uint32 {
-	w := uint32(maxInt(workers, 1))
-	if maxW := mapCap / heartbeatSlotsPerWorker; w > maxW {
+	w := maxInt(workers, 1)
+	if maxW := int(mapCap / heartbeatSlotsPerWorker); w > maxW {
 		w = maxW
 	}
-	return w * heartbeatSlotsPerWorker
+	return uint32(w) * heartbeatSlotsPerWorker
 }
 
 func maxInt(a, b int) int {
